requirements.txt
anthropic>=0.40.0
typer[all]>=0.12.0
rich>=13.0.0
watchdog>=4.0.0
PyYAML>=6.0
PyMuPDF>=1.23.0
Pillow>=10.0.0
python-magic>=0.4.27
rumps>=0.4.0
